refactor: replace ?tab= views with dedicated routes
- /athletes/form-questions for the registration questions editor
- /volunteers (roster), /volunteers/shifts, /volunteers/judges,
  /volunteers/signup-questions, each with a scoped loader
- judges route redirects online competitions to the roster
- sidebar active state uses longest-prefix path matching
- breadcrumb labels for new segments
